Align examples by adding Model folders
