+++ source/ProcessManager.Example.Orchestrations.Abstrations/Processes/BRS_X02/NotifyOrchestrationInstanceExample/V1/Model/NotifyOrchestrationInstanceExampleNotifyEventV1.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model;
 
-namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X02.NotifyOrchestrationInstanceExample.V1;
+namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X02.NotifyOrchestrationInstanceExample.V1.Model;
 
 /// <summary>
 /// The notify event for BRS_X02_NotifyOrchestrationInstanceExample V1.
