+++ source/ProcessManager.Example.Orchestrations.Abstrations/Processes/BRS_X02/NotifyOrchestrationInstanceExample/V1/Model/StartNotifyOrchestrationInstanceExampleCommandV1.cs
@@ -15,7 +15,7 @@
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model;
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model.OrchestrationInstance;
 
-namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X02.NotifyOrchestrationInstanceExample.V1;
+namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X02.NotifyOrchestrationInstanceExample.V1.Model;
 
 /// <summary>
 /// The command starting a BRS_X02_NotifyOrchestrationInstanceExample V1.
